+++ src/test/spec/unified_runner/entity.rs
@@ -49,6 +49,14 @@ pub(crate) enum Entity {
     ClientEncryption(Arc<crate::client_encryption::ClientEncryption>),
     None,
 }
+
+#[cfg(feature = "csfle")]
+impl std::fmt::Debug for crate::client_encryption::ClientEncryption {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("ClientEncryption").finish()
+    }
+}
+
 #[derive(Clone, Debug)]
 pub(crate) struct ClientEntity {
     /// This is None if a `close` operation has been executed for this entity.
@@ -210,7 +218,7 @@ impl ClientEntity {
         }
     }
 
-    #[allow(dead_code)]
+    #[cfg(feature = "csfle")]
     pub(crate) fn client(&self) -> Option<&Client> {
         self.client.as_ref()
     }
